Added new game "Hangman"
The new game Hangman has been added to Language Practice. In it, a definition is shown and a word with all their letters changed by underscores. The user give letters and, if they are in the word, the underscores are changed by these letters. If the user gives a letter that isn't in the word, a life is lost (initially the usar has 5 lifes). If the user gives the correct word, wins the game.